Compression Streams
Decided to look into the browser support for the Compression Streams API again, and turns out Safari recently added support for it in the Technology Preview! It was back in August, I didn't even hear about it.

My existing code was structured so it can work with the Stardazed Streams Polyfill, as this version doesn't seem to work with that for some reason.

Now that there's news of Safari soon natively supporting the API, I'm just going to move my library fully over to that ahead of time. NBTify isn't quite ready for mainstream app use/embedding yet, so not having Safari support for decompression at the moment doesn't seem to bad. Can't wait for Safari's support! It's already working over in Firefox too, so that's great. *Edit: Not sure where I read that, looks like it's only in Chromium still :(

https://developer.chrome.com/blog/compression-streams-api/
https://webkit.org/blog/13137/release-notes-for-safari-technology-preview-152/
https://github.com/WICG/compression/blob/main/explainer.md

*Edit edit: Welp, I think it's not quite cross browser yet, unfortunately. I think I may make my own polyfill for it though, as I want to have easy ESM and TypeScript support for it.

*Aah! Ok, just did some more searching, and found fflate, and it's exactly what I'd want XD; Definitely gonna check it out.
https://www.npmjs.com/package/fflate

Still gonna go through with these changes, I've been wanting to tidy up the compression module for some time now :) I'll find a nice solution in the end, I always want cross-browser and cross-platform support!