toStringTag as Const
Super neat find! I didn't make this connection at first, it's really clever. Turns out you can force a function return type by using `as const`, and it will apply that to the implicitly return value! So if you always return `5` per say, then you can declare it with `as const`, and it will force the return type to be `5` specifically, rather than `number`! This is super cool, since that means you only have to declare it in one place :D