NBTData Rewrite + Option Types + Coding Bonanza
Rewrote the `NBTData` object to be way more secure like the rest of the library! Now it specifically allows only certain data in, and it will be frozen immediately after it is constructed. I want it to work like a native platform class would, so I have been following how `ImageData` works, and what it's data flow is, and how secure it is.

When `ImageData` is constructed, all of the resulting properties on the constructed object are all readonly, and you can't add any more properties to the object either. Thanks yet again to StackOverflow, turns out this is a great use case for `Object.freeze()`! It's also a bonus that I *don't* want the `NBTData.data` property to readonly, that one should be editable. `Object.freeze()` only freezes the object your pass into it, so it won't recursively freeze objects that are inside of that one. Sometimes that would be what your are looking for, but I don't really need that for this case.

https://stackoverflow.com/questions/67698353/is-there-an-elegant-way-to-set-a-class-property-only-once

Along with the readonly-nature of `NBTData`, it now also has error handling! So if you pass in broken parameters into the constructor, it will get you back for it! *No more rotten data for me, no thank you! - NBTData*.

For the option types, they now have their own separate dedicated type definitions! No longer is there a general `Metadata` interface which has those types on it. Now you can way more easily pick and choose which ones you need, and add those to a new interface you're making.

Oh yeah! I also got recursive `new NBTData()` constructors working too! So, since `NBTData` objects will now be readonly, there wouldn't be a way to easily make a new object with that same data. Before you could just modify the property on the object yourself, then do what you will, but now that's not available anymore. Instead, you call it like this: `new NBTData(new NBTData({ thing: "noice" },{ endian: "big" }),{ endian: "little", compression: "gzip" })`. You probably wouldn't need to do it directly like that, it's more for transforming an existing set of NBT data that you loaded in, and you may want to convert it to a different format type, say to change the endian type. I for example will eventually use this to convert a set of NBT data from one Minecraft version to another.

Reworded some of the error text in a few spots, mostly generalizing them to be less wordy, and changing `"argument"` to `"parameter"` (Noticed I was doing that somewhere else, and I think I like that better :D ).

The constructed `NBTData` object isn't actually quite set to readonly yet, as there's a little bit of a messy spot in the `read()` function that requires it to be editable, so I left that for after this update. I wanted to safely push these changes while they definitely work, before I screw things up XD

Eyooo, I don't know how I've been coding so much this week, it's been really refreshing! Changing between a lot of my different projects, and taking breaks when I work on it for too long has been helping out a *TON*. I've been wanting to get essentially these set of changes to the `NBTData` object going for a while now, but sometimes it just hasn't clicked yet. I think it's when you don't quite know how you want it to work yet, and that kind of holds you back. Somehow, going to a different project and working on that instead really helps you for when you come back to the original project. You learn one thing over in the second project, and then it unexpectedly helps with another mountain in the first project! It just clicks sometimes, for some reason. This many commits this month is crazy, I think it said 55 so far, if I remember right. I don't know if I've had that many in a month before! I'm fine if that means I need to take a break after this crazy brainstorm. It's been a nice ride of tackling thing after thing, and I just keep wanting to see what other little things I can conquer for each project. Writing things like this has been very nice for me too, I essentially do this over on Strava too much too haha, pretty much just as long as this one is here. I think I definitely will look into making a journal entry loader kind of thing, which would dynamically turn my Git commit messages into a journal with dates, of some sort. It would be like an auto-blog, driven by my Git history XD. It would have dates and all those kinds of things too, so it would look like a news site, but it would be with articles like these, or whatever this whole messages is. Essentially, a code-driven life update! Aaah, this is so cool. O-PEN-SOURCE! O-PEN-SOURCE! O-PEN-SOURCE!

Listened to Rush Counterparts on the way home from work last night, it's mixed so well! Same with the new(ish) Dream Theater (AVFTTOTW). You can turn them up loud, and the speakers don't start to die on you even with it being loud.

I have noticed an interesting correlation between how what I think the creative part of my brain performs. During this coding bonanza the last few weeks, I feel like I haven't mentally had musical energy to write any songs, which is interesting. I think music and coding probably have a similar room in my mind (that should be a song!!! *Similar Room in my Mind*), so sometimes one takes up more space than the other. After this coding wave goes down, I semi-predict that my music drive will be back strong like the coding drive has been. It's not that I haven't wanted to work on music, but I recorded a few songs, and I felt like it was the same-old same-old, and that kind of turned me away. I know well by now that not all songs are great (I love that!), so it's not that the song was that bad. It's like I didn't want to try and fix it to where it should have gone to. Just like the coding projects, I think leaving it to dry out for a little bit helps freshen it up, more so than trying to bang it with a hammer until it doesn't even fit in the round hole XD

Ok, I think that was everything I was gonna cover? Ya never know haha. Gn!