Empty Lists: Part Two
Got empty list writing working! Did the initial testing of it with `simple_house.nbt`. That appears to have a different parsing issue though, which is corrupted on the rewrite after reading. So, I made a supported NBT file with NBTify which had a single Double entry in the list, saved that, edited it with WebNBT to remove the entry manually, then worked with that to add the fix to NBTify to allow it to write empty lists too! This currently follows the Notchian spec for empty lists, as defined at https://wiki.vg/NBT. If an empty list has a type when read initially, that is unfortunately lost when reexported to raw NBT again. But, I think that probably happens with the regular implementation too (I'm not totally sure about that though), so I'm fine with that loss, as it appears to be an expected outcome that the spec outlines.

This does fix the empty list issue, but yeah, something is still a bit wonky with `simple_house.nbt`. Opening and reexporting that with WebNBT kindof works, but it reeorders all of the compound keys, which is harder to test the changes against. I'm glad that my library retains the original key order!

Now I want to go through a ton of my GitHub issues, as a ton of them are supported/fixed now! Super cool! I can't believe how strong the typing is for all of this too, thanks to TS :)

Gonna look into being able to write a plain CompoundTag directly using `NBT.write()`, since right now it currently only accepts an `NBTData` object, which I found to be a hassle when quickly making a new demo for the empty list testing.

Lunatic Soul makes me happy, this is a very cool album, Through Shaded Woods :)
Listening to it twice in a row, aaaah :)

Edit: Oh yeah, empty list *reading* had no problem, from what I could tell, it appeared to be only the writing part. My implementation infers the list type by getting the NBT type of the first item in the list. In an empty list, that will be `undefined`, and that's not a valid NBT value. That would in return cause the `getType()` function to throw with an error that `undefined` is not a valid NBT type. My fix uses a ternary to set the type as TAG_End if there is no first item in the array. I thought about using `list.find(entry => entry !== undefined)` to get the first non-`undefined` item in the array, and using that for the type, but that's an unneccesary extra step, since the `undefined` values you are skipping over in the first place are simply going to be unserializeable to NBT anyways, so it's not like they can be used anyway. So, if the first item is invalid, then that means the whole list is invalid, because no matter where an undefined item is in the array, it's invalid! I don't want to filter out `undefined` values and serialize everything else, as that's kind of over-correcting the issue, and I want it to be statically type-strong, if that means anything haha. I do still have to make a check for whether all of the list items are of the correct type that match the first item though, currently it doesn't do that on the implementation side of things. If you list has multiple item types, that is not currently secure at the moment. It should definitely throw an error if the list has multiple item types, as you are only allowed to do one item type for a list. Dang, that was a very long note haha. Also, the 2nd time around for the album just finished. Great timing :D

Edit2: Oh yeah, oh yeah, I reedited `getType()` to no longer accept any value. I was going to use it both to check whether a value is a valid type or not, and to get the tag for that if it is a valid type. I think this is a bit convoluted though, as it should be a different function for that I think, as this one should be tailored mainly to get the tag for a *valid* value. Having a separate `isTag(value: any)` kind of thing could give a boolean for if a value is a supported tag type. APIs are messy, ok?? haha. It's getting better though.