Import/Export Organization Pt. 2 + Type Flattening
Adjusted some of the imports and exports to make things more descriptive, like using the `type` modifier on `type`-only imports, and re-organizing the `index.ts` exports to be organized by abstractivity (like specificity, but for abstractions' sake) rather than alphabetically. I think this is the same order that I used for them originally.