Read-Only NBTData + Proper Recursive Constructor
Re-added the `BedrockLevel` type, and I will actually probably end up doing that for the other types again too. It's nicer to have them all under one name, and it's more apparent to the user of the library what the value will expect if it's listed as an export.

Marked/made all of the `NBTData` properties as readonly/frozen! Part of being able to add that, now you can recursively create new `NBTData` objects using the constructor, and it will auto-populate pre-existing properties from the previous object, and it will override any of the properties you redefine in the new constructor's options field. More prepared for now, you can also remove any of the optional properties from the new object by passing `null` as the value for the property in the new constructor. So, if you had `bedrockLevel` set on the old `NBTData` object, you could pass the old `NBTData` object as the `data` parameter, then set `bedrockLevel: null` in the options parameter. Super neat! Then that will return you an equivalent `NBTData` object, with the only change being the removed `bedrockLevel` entry. And both of them are still readonly! I use `Object.freeze()` to lock the properties. Thanks to StackOverflow for reminding me about that one! I was initially going to use `Object.defineProperty()` for each of the properties when I wrote the `NBTData` class, and `Object.freeze()` is much simpler. Note: Decided to try this on an array, since an array is just an object too, and it works! `[].push()` throws when calling it on a frozen array! Definitely great to know for later.

With all of the behind-the-scenes information I've learned about JS, like with the prototype chain, things like that, it's really great to get how an array in JS is really not an array, but an object with numbers for the keys, and that's how it stores the items. Makes perfect sense as to how they allow any value, it's just a plain object, that's why! Super cool actually, and I'm glad it's like that. You can do so many cool things if you embrace how JS is setup to work. Yes, it may be wacky compared to some languages, but the dynamic nature of it is *very* expandable! I love you, JS!

Lunatic Soul - Through Shaded Woods, listen v5 I think :)