Unnamed Root Compound Tags
Added initial support for unnamed root compound tags! This currently works at the `NBTReader` level, and fully at the writing level. Because of all the new additions to the `NBTData` class, the composition functions (`read()` and `write()`) could use a little help to better reflect those new updates. One example being, also accepting either simply just a `CompoundTag`, or an `NBTData` object, and also moving all of the validation logic over to the `NBTData` object instead of replicating it in all of the parameter input locations.

Big thanks to WebNBT for being able to deduce the format for this! Looks like not even Prismarine NBT supports unnamed root Compound tags. nbt-ts, does support it though! I checked out their support for how they got it to work. And the format simply just removes an initial `StringTag` after the first `CompoundTag` byte at the start of the file. So, minus two bytes. I'm gonna have to add a check for that in the file type deduction code inside of the `read()` function, similar to how the endianness check currently works. I'll need another check that will also perform a root-name-or-not kind of thing. Glad to have all of the `NBTData` properties declared safely now, relying on those for parameter management will make things much better for the future IO reworks.

Re-added the `Name` type, since it is now more complex than just a `string`, now also possibly `null`. Compared to the optional properties on the `NBTData` object, `name` is not optional, but rather will be listed with a `null` name property. So, it is either `string | null`, instead of `PropType | undefined`.

Getting sleepy, listening to Yellow Hedgerow Dreamscape now. Gotta get up early for my potato buckets shift in the morning, so this will likely be my closing commit for the night. zzz, gn! :)