List Length Spec
Changed all of the list types that were out of compliance with the spec on https://wiki.vg/NBT. Other than the TAG_String type, all of the length prefixes for tags should be their signed equivalents. I previously used unsigned for all of the length types. I thought that would allow for a longer list type, which would be great, but maybe I'm missing something? Thinking about it now, I wonder if that's because of some other numerical limitations in the format, and that forces the lengths to be under that certain size. I thought that having it be unsigned cuts the (positive) numberical avaliablility in half, right? Not totally sure, but I want the reader and writer to be spec-compliant rather than improving the format, so I'm gonna go with that instead actually. Curious about that last part though, I wonder if you could make the list tag length longer if the length bytes were unsigned? :O