Buffer Offset Fix
Somehow haven't checked the Bedrock Level Dat test file in a little bit, and this totally broke because of the `DataView` offset change. This fixes that, and adds addition math to account for when a `Uint8Array` of NBT data isn't at the same byte offset or byte length as it's parent buffer. Super neat! Also, still experimental. I may add back the `#data` internal property, as that makes calculations much easier to find the offsets and size of thing when reading the buffer.

This was a funny issue, as it still was able to parse the data correctly, but the offset for the tag names were consistently off by the `byteOffset` property, so they looked super wonky and garbled, but it was actually parsing as expected XD