Smart Buffers!
Expanding off of yesterday's find and solution, this update make the library's buffer handling a bit smarter. Now that I know how `TypedArray`s interact with `ArrayBuffer`s a bit better, I made it so original `ArrayBuffer` input data can be used as much as possible. I didn't know how to work with the `byteOffset` and `byteLength` properties completely yet, and now I know how to work with them, so I'm using that to make new `TypedArray`s that reference the same underlying `ArrayBuffer` data directly, rather than duplicating certain sections of it just for reading purposes. I haven't visibly noticed any performance changes yet, but this will definitely help with more hefty files eventually, once I get to working with those. It sounds much more performant to read the data at a specific offset, rather than duplicating the same data just to have a certain section of it.

This brings back the internal `#data` property to the IO classes, as it makes the data access and calculations much easier. The `Uint8Array` itself now manages where you read the underlying data from, rather than the library manually applying that to the `DataView` and `.slice()` calls.

These changes are mostly to do with the reading side of the library, as the writing side won't have any custom `byteOffset` or `byteLength` properties set, because the library can manage the length of the buffer as it needs to, with `#accommodate()`.

After a little bit of Googling, and reading, this article covers how this works *very* nicely, and it helped confirm what I thought I now understand. Yay!
http://shockry.blogspot.com/2017/04/javascript-typed-arrays-slice-vs.html

Turns out I was using `.slice()` wrong too, at least in terms of performantly reading the data. `.subarray()` does the same things as before, but now even better because it uses the exact same data, rather than making a copy of the base `ArrayBuffer` between the indices!

Not in this update, but I'm going to look into it next, I want to redo/verify how the `#accommodate()` function works, just to see if I can make it more performant or straightforward too. I found this post which covers that other implemenetations tend to double the length of the buffer when it needs to expand, and that writing to the buffer performs "in O(1) amortized time". I'm no expert on O-notation yet, but I think that's the good one you want to have, hehe. I'm definitely curious what the different trade-offs are in terms of how often to expand the buffer (make a new one at x length), vs making the buffer larger to prevent having to expand it as often. It seems like from the description here (and thingking like O-notation) that having to reallocate less often is a nice sweet spot, even though it is more space. Definitely gonna look into that! :D
https://stackoverflow.com/questions/45898229/recommended-way-to-grow-a-buffer/46138231#46138231

Oh yeah, this is a semi revert/rework of the changes made in "IO Class State Properties". That's when I removed the internal `#data` properties in favor of just using the raw `ArrayBuffer` from the view instead. Now I have a `TypedArray` to access the data from the correct offsets, as expected!

I'm glad that the Bedrock Level file kinda broke this, because I've wanted to understand how this ArrayView setup has worked, since essentially the beginning of this whole project. I think I've understood the idea of what it was doing, the view vs the raw data itself, it was just the implementation of managing that kind of setup that was tricky for me. Now that I have a working version that manages it, it's super neat! `TypedArray`s are cool :)