Explicit Read Options
Added support for explicitly disabling specific read properties as needed. Previously, the `read()` function would always just smart it's way around getting the file to open, but I want to also allow you to try and open the data as you need to. If you want to try and specifically open a file using big endian, and with a Bedrock header, why not? Not sure what kind of wackiness is out there, so this allows you to be more explicit with what exactly you're trying to open. I think this could be helpful if you're trying to figure out the format of an existing NBT file you may have, and to where NBTify isn't able to deduce it's format for some reason. More config, the better!

This first version for these additional options initially deal with declaring no compression is used, and that a Bedrock header check should/n't be performed or not. Next one to work on will be the `isNamed` option, which will be replacing `named`. I'm using the `is-*` prefix since that will better represent that you are passing a boolean, rather than a value, unlike the options for the `NBTData` constructor, and `write()` function, which accept the values themselves, which will then in return distinguish which format settings you'd like to use.

TypeScript also really helped out with this a lot. I also tried to do this a while back, but things got too messy/confusing. I left it to the auto setting (which works great on it's own) until I could figure it out exactly. Another big part is deciding to use `null` to represent where you specifically don't want a config option to be set. `undefined` is essentially reserved for what auto/smart would be, and `null` is for where you don't want to use any of the option values, or the auto/smart setting. This `null` behavior is only applicable to optional properties, like `compression` or `isBedrockLevel`, whereas `endian` has to be either `big`, `little`, or auto/smart. That one has to be provided one way or another.

This also fixes being able to pass in `zlib` as a compression format, and the `read()` function still auto-reading as `gzip`. Yeah, that was super wonky. `zlib` has merely been a placeholder for the whole time so far though, so it's not breaking or anything for it to not have worked in the first place. Still weird though. So now `compression` has to be explicity set to `gzip` or left as `undefined` to decompress with gzip, and it won't work with anything else (`zlib` or `null`).