isNamed Option: Part Two
This adds the actual implementation for the `isNamed` handling. It also includes the auto handling too :O

That Never Nester technique is working wonders on this function. These optional property code block handling segments work amazing! I didn't even see that pattern until I did the Never Nester trick on it. This essentially makes it so the rest of the function after the checks are marked as safe (non-`undefined`), since they return early from the function with the correct value! Outstandingly neat!!! That happened to work out that way, I'm not sure if I could've planned that structure. I was gonna make a snippet to describe it to you to see how it works, but I'm not gonna make a demo as close to the actual code that shows how it works, haha. I tried to make it, but it's just the same code again. So the island exit early lines are on lines 35 and 49 of `read.ts`. They will deduce the property type of the file using `try {} catch {}` blocks, and then they will re-call the same function again, recursively until one of the types passes! (That description actually turned out really nice) This in result causes everything after the recursive-call blocks to then narrow the parameter type to remove `undefined`, because you already `return`ed from that initial function call, where it wasn't declared! It's then in that next function call where that parameter has now been deduced, and is no longer `undefined`. Wow. Wow, wow, wow, wow, WOW! Maybe there's a graph-like tool in Chrome DevTools that can draw out the function stack so you can see how it runs? That would probably help explain it too.

This is sick, I love that this is a thing. Oh yeah, this project is turning a year old in about a week. Pog XD

*Edit: Oh yeah, I was gonna initially try to add the `isNamed` handling further into the `endian` try-catch loop, but this setup renders that totally unnecessary! I'm super happy about that, because that would've without question been a spaghetti code nightmare with even more nested try-catch statements. I already don't like how many try-catch nestings there are, and that would've made it a lot worse. I may look into adding `.catch()` statements to the end of the `await` calls, luckily since the `read()` function is async. I know how to use `.then()` nicely now, but I haven't done much with `.catch()` yet. Guess there's no reason to `await` around, HAHA! I think that may take care of some of the try-catch nesting hopefully. That would look amazing as simple callbacks :)