Elimination Deduction
Redid the format deduction handling, and it's super cool! It essentially brings the setup from the other format handling features, and brings it to the `compression` and `isBedrockLevel` properties also. The idea is to eliminate/deduce "auto" format types as soon as possible, because you can then use that found format to find the ones that follow. So the flow is now that it will determine if the file is compressed with gzip, it will iterate-loop through all of these values with both the `big` and `little` endian types until one of them works (this eliminates the `auto` because you are essentially restarting from scratch with a value to test with, either `big` or `little`), then it will determine if the file has a Bedrock Level header (requires the file to be `little` endian. If the file has the header, and it tries to read it as `big` endian, it will then fail to parse, and then try the `little` endian type, which will then work, and it is able to fully check for the header since `little` endian is now the endian type to check), next it will iterate-loop yet again for the `isNamed` property, first trying with `true` (same flow as `endian`, this time try-catching with attempting to read with a root name, then without). Now that all of the properties are found, it will "try to parse the data" (it has already done it a few times to discover the properties, but now this is the one that will work), first by decompressing by the specified format (if one was discovered), then slicing off the Bedrock Level header (if it was marked as being present). Then, you have your `NBTData` object! Very awesome.

So, I think that means that the most complex file to parse would be one like this:
```js
{
  name: null,
  endian: "little",
  compression: "gzip",
  bedrockLevel: new Int(8) // Any Bedrock Level version number
}
```

I haven't tried to make a file like this yet, but I think this would need to pass through all of the auto-checks to end up parsing out. I think the opposite would simply be a named, uncompressed, big endian, non-Bedrock Level file. By default the reader will read for a root name present, big endian, no compression, and without a Bedrock Level header. So, essentially, `hello_world.nbt`'s format! After this, I'm gonna have to make a demo file with all of those crazy critera, haha! If that ends up working, I'ma be stoked! Let's see...