Ridiculous NBT
Alright! Made the demo file with the crazy critera I was talking about in the last commit. After looking into a bit more of some info about the Bedrock NBT format (over on https://formats.kaitai.io/minecraft_nbt/, the wiki didn't really specify this), if a Bedrock Level header-applied file is compressed with gzip or zlib (in terms of the game, and what it will do), the header is part of the compressed data, rather than being applied after the compression step. When writing to the raw buffer, I was first applying the compression to the data, then adding the header after that. Now I have it flipped around to what it sounds like it should be instead, which is applying the header to the uncompressed raw NBT data, then bundling all of that into the compression format together. Once making that change, I generated my custom `ridiculous.nbt` file with all of the ridiculous criteria, haha. Here's the code snippet I made to generate that:
```js
const value = {
  IsRidiculous: new Byte(1),
  Bugrock: new Int(24),
  Noice: new Int8Array([7,3,9,6,1,5,4,3,1,33])
};

const result = new NBTData(value,{
  name: null,
  endian: "little",
  compression: "gzip",
  bedrockLevel: new Int(8)
});
```