Buffer API Documentation
Changed references to using `Uint8Array` objects as IO for the library, over to a more generic "buffer" terminology. As noted in one of the repo issues/goals, I want to allow the user to be able to open both `Uint8Array`/NodeJS `Buffer` objects *and* plain `ArrayBuffer` objects. This makes it more general in terms of what the API can accept and work with, once I get to adding `ArrayBuffer` support.

I also shortened some extra descriptions that were on the IO classes, and instead reworked their parent IO functions, `read()` and `write()`, to make those more straightforward. I plan on having the IO functions be the main entry point to where you would want to use the library, and the IO classes be a public way to see how it works, or if you needed to do something under the hood with NBTify that I don't know about ahead of time. Never know what someone might want to make with your API! Having as much available to them to do custom things seems like the cool/helpful thing to do :)