Explicit Write Options
Essentially the inverse of "Explicit Read Options", for the Read module, this brings the same things over to the Write module. Now you can disable specific format types when writing to an NBT buffer (`Uint8Array`). This means you won't have to create a whole new `NBTData` object if you just wanted to one-off write an existing NBT tree to a buffer, using a different set of formats. Now you can just pass those in as options!

I also added an additional check against `undefined` and `null` in the Tag module's `getTagType()` function. Now it won't try and validate those as a CompoundTag, as that used to show the incorrect error before. I've been trying a few ways of improving the "... is not a valid NBT primitive" message, but haven't been happy with any of them yet. That will be for another update.

Back to the Write module, I removed the extra try-catch block around the writing process, as it wasn't doing anything helpful to the error handling. I also found an article that says that that is a code-smell kind of thing haha, so that made it seem not neccessary to have there.

Also added some missing parameter validation to the `write()` function. Looks like I didn't get around to adding that when I expanded the format options a while back.

Oooh, a real nice one! While trying to test the new `write()` param handling out (for nullifying `bedrockLevel` on the base `NBTData` object), I discovered that the auto-reading inside of `read()` is returning early if the data passes as big endian, when reading the Bedrock Level (.dat) file without a Bedrock Level Header. This is because of the missing checksum length feature/bug, and I think that may have caused some of my issues before without me noticing. This is a big one for the auto-format detection! This breaks it from working altogether. I'm glad I found an example that I can use for it which I can consistently test on. My fix for the test, for this update at least, is to ensure that the data be read as little endian, as it was passing as big endian with the auto-format detection, even with still the whole buffer of bytes left to read after that.

So, the buffer length checksum feature/fix will be soon on my agenda for things to come also, as that will help fix a lot of possible bugs that may arise from trying to read any mysterious NBT!