Bytes Read Success Checksum
Added the bytes remaining check to `NBTReader`! Inspired from Prismarine NBT, this protects against early reading finishing, when the reader encounters an End tag-like byte, based on the given format it's trying to parse. Say if you tried to read a root-named NBT data tree (with `""` as the name) as an unnamed one, it would encounter the String tag length of `0` and treat that as the End tag for the root Compound, and complete the reading process with no data in it. This would be an incorrect "error-free ending", as it didn't actually read until it found something unexepected. It actualy read until it was told to end, unexpectedly, haha. This makes the auto-format deduction must more tight, since the root-name-or-not handling would cause this to happen when auto-reading a non-Bedrock Level little endian file. There's other instances of this, where I have noticed a similar error, but hadn't yet realized it was this bug that was causing it.

I also want to add nicer error handling for when `NBTReader` is trying to read past the end of the buffer. Right now it mentions things about it's internal `DataView`, and the `byteOffset` it's trying to read at, but that's not so apparent to the user that that is what's going wrong, based on the error. At least from my experience, I didn't figure out that's what it meant until it only just clicked a little bit ago. Just made a new issue for this over on the repo page, while writing this paragraph :)